Refactor: CommunityCreatePage -> CommunityEditPage로 이름 변경
