--- src/pages/CommunityCreatePage.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import CommunityCreate from '../components/community/CommunityCreate';
-
-const CommunityCreatePage = () => {
-  return <CommunityCreate />;
-};
-
-export default CommunityCreatePage;
