+++ src/pages/CommunityEditPage.tsx
@@ -1,8 +1,8 @@
-import React from 'react';
-import styled from 'styled-components';
-import CommunityEdit from './community-edit/CommunityEdit';
+import React from "react";
+import styled from "styled-components";
+import CommunityEdit from "../components/community/community-edit/CommunityEdit";
 
-const CommunityCreate = () => {
+const CommunityEditPage = () => {
   return (
     <Wrapper>
       <Header>
@@ -14,7 +14,7 @@ const CommunityCreate = () => {
   );
 };
 
-export default CommunityCreate;
+export default CommunityEditPage;
 
 const Wrapper = styled.div`
   width: calc(100vw - 300px);
@@ -28,7 +28,6 @@ const Wrapper = styled.div`
     width: calc(100vw - 100px);
     padding: 3rem;
   }
-
   ${(props) => props.theme.response.mobile} {
     width: 100%;
     padding: 2rem;
@@ -40,19 +39,15 @@ const Header = styled.div`
 
   > div:first-child {
     margin-bottom: 1rem;
-
     font-size: 50px;
     font-weight: 700;
     color: ${(props) => props.theme.color.primary};
-
     ${(props) => props.theme.response.mobile} {
       font-size: 45px;
     }
   }
-
   > div:nth-child(2) {
     font-size: 20px;
-
     ${(props) => props.theme.response.mobile} {
       font-size: 17px;
     }
